autogen-mcp-system/requirements.txt
# FastAPI & Web
fastapi==0.115.7
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1

# MCP Framework
fastmcp==2.13.0.1

# AutoGen 2
autogen-core==0.5.7
autogen-agentchat==0.5.7

# Ollama Integration (no LiteLLM needed for 0.4.8+)
ollama==0.5.3

# Microsoft SQL Server
pyodbc==5.2.0
sqlalchemy>=2.0

# LDAP Authentication
ldap3==2.9.1

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Configuration & Secrets
python-dotenv==1.1.1

# Utilities
requests==2.32.5
tenacity==8.5.0

# Logging
loguru==0.7.3